requirements.txt
pipx
nb-cli
nonebot2[httpx]
nonebot2[websockets]
git+https://github.com/EliwiiKeeya/nonebot-adapter-discord.git@master

ujson
pydub
qrcode
pillow
opencc
pymongo
opencv-python
